+++ src/main/java/esa/sen2vm/input/OptionManager.java
@@ -241,7 +241,7 @@ public class OptionManager
            
             Option steps = new Option(OPT_STEP_SHORT, OPT_STEP_LONG, true, "!!! MANDATORY !!!: Steps (pixels) for bands 10, 20 and 60m \n"
                                         + "(separated with whitespace, respect the order)");
-            steps.setType(Float.class); // TODO  does not seem to work: read as array of String 
+            steps.setType(Double.class); // TODO  does not seem to work: read as array of String
             steps.setArgs(3);
             steps.setRequired(true);
            
@@ -270,7 +270,7 @@ public class OptionManager
 
             // For inverse location: the following options are compulsory
             Option ulxOption = new Option(OPT_ULX_SHORT, true, "(Mandatory for inverse loc) Upper Left X (referential unit)");
-            // TODO ulxOption.setType(Float.class) does not seem to work: read as String 
+            // TODO ulxOption.setType(Double.class) does not seem to work: read as String
             ulxOption.setRequired(false);
             Option ulyOption = new Option(OPT_ULY_SHORT, true, "(Mandatory for inverse loc) Upper Left Y (referential unit)");
             ulyOption.setRequired(false);
@@ -286,19 +286,19 @@ public class OptionManager
             outputFolderOption.setRequired(false);
 
             Option outputImageRes = new Option(OPT_OUTPUT_IMAGE_RES_SHORT, OPT_OUTPUT_IMAGE_RES_LONG, true, "(Mandatory for inverse loc) output folder");
-            outputImageRes.setType(Float.class); // TODO  does not seem to work: read as array of String
+            outputImageRes.setType(Double.class); // TODO  does not seem to work: read as array of String
             outputImageRes.setArgs(3);
             outputImageRes.setRequired(true);
      
             // params arguments
             // ------------------
             Option detectors = new Option(OPT_DETECTORS_LIST_SHORT, OPT_DETECTORS_LIST_LONG, true, "(optional) List of detectors to process separated by spaces, example: 01 05 06 10 11");
-            detectors.setType(Float.class);
+            detectors.setType(Double.class);
             detectors.setArgs(Option.UNLIMITED_VALUES);
             detectors.setRequired(false);
 
             Option bands = new Option(OPT_BANDS_LIST_SHORT, OPT_BANDS_LIST_LONG, true, "(optional) List of bands to process separated by spaces, example: B01 B08 B8A B10 B11");
-            bands.setType(Float.class);
+            bands.setType(Double.class);
             bands.setArgs(Option.UNLIMITED_VALUES);
             bands.setRequired(false);
